feat(ui): unify shield screens into single ShieldScreen with address selection
Merge ShieldCreditsScreen and ShieldFromAssetLockScreen into a single
ShieldScreen that uses AddressInput to let users pick the source. The
screen adapts its behavior based on the selected address type:
platform addresses use ShieldCredits, core addresses use
ShieldFromAssetLock. One button in the shielded tab replaces two.
